Parse ENUMERATED with explicit variant numbers and default markers #6
This parses ENUMERATED with variants that have explicit variant numbers
and default markers (...) but it does not reflect those new information
in Rust, Protobuf or SQL yet. It just allows them to be parsed without
failing the parser. Trying to generate Rust, Protobuf or SQL will result
in a panic.