Parse tags (ignored in Rust/Protobuf/SQL) #6
Although UPER is used - which does not care about tags at all - protobuf
is using tags. Maybe it will be useful in the future, to remember the tags
when converting a Model<Asn> to a Model<Rust> or Model<Protobuf> - i.e. if
one is going to implement a BER or DER (de-)serializer.