Add draft implementation of a new type representation to reduce complexity
The current approach to generate the (de-)serialization code is to messy,
too hard to understand, too hard to extend and too complex. The new idea is
to introduce a further codegen step.

1) ASN Definition   -- old codegen -->  3) Rust struct with (de-)serializable

   |                                                    A
   | generate                                proc-macro |
   V                                                    |

 2) Rust struct with attributes and that pleasant for human eyes

Instead of writing 3) to the user file, 2) shall now be written instead.
Also, instead of generate comple (de-)serializ code in 3), the Visitor
pattern is being used to walk the types and values on the generated data
types. This hopefully allows to implement UPER, PER, BER, ... #1 #10 #11