Add common::Constraint which always provides a Tag for the current type, fixes #13
Resolve all tags while augmenting the annotated structs with Constraint impls. It
is neccessary that to this state, the complex-type-references have already resolved
their tags. For this it it might be neccessary from now on to have all *.asn1 files
and their modules in scope while resolving the tags in the RustModel-to-Annotation
stage.

Quite a bit of work went into ensuring that no unncessary tags (meaning default tags)
are present in the annotated representation to prevent unnecessary bloat.

This will probably allow the SET and SET_OF branch to be merged soon #11 #10