Breaking change: split Error and box it, add "descriptive-deserialize-errors"
Boxing the error provides an performance uplift of about 10% and provides a
place to store the ScopeDescription list intorduced by the new feature
"descriptive-deserialize-errors". If it is enabled, it collects various
descriptions of internal state changes (including the names of the currently
deserialized types) and thus allows one locate the origin of deserialization
errors.